The Second Standardization Update, Part I - Add "common.h", and update templates (minus MPI currently) and programs to use common functions without needing so many externs - Also add template support for arbitrary input seedlists and/or output files if desired - Add TL;DR to license - Partially organize Extreme Climates list - Add Perlin Noise percentiles to Utilities, to speed up programs at the cost of some accuracy - Update Pybiomes

+++ Far Spawns/1.18+/Perlin Octave Bounds/(Unfinished) shgo.c
@@ -9,7 +9,7 @@ typedef struct {
 	double first, second;
 } Pair;
 
-void* shgo(void* func, Pair *bounds, size_t boundsCount, void **args, void **constraints, int n, int iters, uint8_t sampling_method) {
+void* shgo(void* func, Pair *bounds, size_t boundsCount, void **args, void **constraints, int n, int iters, int sampling_method) {
 	if (sampling_method >= NUM_SAMPLING_METHODS) {
 		fprintf(stderr, "shgo: Invalid sampling_method \"%" PRIu8 "\".\n", sampling_method);
 		exit(1);
